+++ src/Interfaces/fonts.ts
@@ -1,0 +1,7 @@
+export enum FontSizeName {
+  XS = 'XS',
+  S = 'S',
+  M = 'M',
+  L = 'L',
+  XL = 'XL',
+}
