+++ src/Interfaces/router.ts
@@ -1,0 +1,13 @@
+export enum RouterPaths {
+  'fonts' = 'fonts',
+  'colors' = 'colors',
+  'elements' = 'elements',
+  'component' = 'component',
+  'exporting' = 'exporting',
+  'assets' = 'assets',
+}
+
+export interface Router {
+  path: RouterPaths
+  componentId?: string
+}
