+++ src/Interfaces/ui.ts
@@ -1,0 +1,15 @@
+export enum DragDirection {
+  N = 'N',
+  NE = 'NE',
+  NW = 'NW',
+  W = 'W',
+  E = 'E',
+  S = 'S',
+  SW = 'SW',
+  SE = 'SE',
+}
+
+export enum ComponentStateMenu {
+  hover = 'hover',
+  focus = 'focus',
+}
