+++ src/editor/LeftMenu/LeftMenu.tsx
@@ -2,7 +2,8 @@ import * as React from 'react'
 import styled from 'styled-components'
 
 import state from '@state'
-import { Alignment, Component, NodeTypes, Overflow, RouterPaths, Units, ViewTypes } from '@src/interfaces'
+import { Alignment, Component, NodeTypes, Overflow } from '@src/interfaces'
+import { RouterPaths } from '@src/interfaces/router'
 
 import AddInput from './AddComponentInput'
 import ComponentItem, { Item } from './ComponentItem'
@@ -11,6 +12,7 @@ import PlusSign from '@components/PlusSign'
 import { uuid } from '@src/utils'
 import { route } from '@src/actions'
 import StaticItem from '@src/editor/LeftMenu/StaticItem'
+import { Units } from '@src/Interfaces/node'
 
 const LeftMenuBox = styled.div`
   box-shadow: rgba(0, 0, 0, 0.12) 2px 2px 2px;
@@ -84,7 +86,6 @@ const addComponent = value => {
   const newComponent: Component = {
     id: newId,
     name: value,
-    viewMode: ViewTypes.SingleCenter,
     root: {
       id: 'rootId',
       type: NodeTypes.Box,
