+++ src/editor/TopBar/TopBar.tsx
@@ -1,18 +1,10 @@
 import * as React from 'react'
 import styled from 'styled-components'
 import state from '@state'
-import {
-  Alignment,
-  Border,
-  BoxShadow,
-  ComponentStateMenu,
-  ComponentView,
-  FontSizeName,
-  NodeTypes,
-  ObjectFit,
-  Overflow,
-} from '@src/interfaces'
+import { Alignment, Border, BoxShadow, ComponentView, NodeTypes, ObjectFit, Overflow } from '@src/interfaces'
 import { Colors } from '@src/styles'
+import { FontSizeName } from '@src/Interfaces/fonts'
+import { ComponentStateMenu } from '@src/Interfaces/ui'
 
 const TopBarBox = styled.div`
   padding: 8px 16px;
