+++ src/selectors.ts
@@ -1,5 +1,5 @@
 import state from '@state'
-import { RouterPaths } from '@src/interfaces'
+import { RouterPaths } from '@src/interfaces/router'
 
 export const getCurrentComponent = () => {
   if (state.ui.router.path === RouterPaths.component) {
